Extract GithubIssueError to its domain module
Extracted GithubIssueError from the centralized error.rs to its domain module at src/run/github_issue.rs. This improves module cohesion by locating error types closer to the code that throws them and reduces the bloat of the top-level error.rs file.

Updated the top-level Error enum and exit_code.rs to correctly reference the new location. Verified that the project successfully compiles and passes tests with these changes.
