+++ src/error.rs
@@ -24,7 +24,7 @@ pub enum Error {
     Github(String),
 
     #[error("github issue: {0}")]
-    GithubIssue(#[from] GithubIssueError),
+    GithubIssue(#[from] crate::run::github_issue::GithubIssueError),
 
     #[error(transparent)]
     Io(#[from] std::io::Error),
@@ -240,19 +240,4 @@ impl From<toml::de::Error> for ConfigError {
     fn from(e: toml::de::Error) -> Self {
         Self::Toml(e.to_string())
     }
-}
-
-#[derive(Debug, Error)]
-pub enum GithubIssueError {
-    #[error("missing GitHub token: set `{0}` to a PAT or GitHub App installation token")]
-    MissingToken(String),
-
-    #[error("GitHub issue or repo not found: {0}")]
-    NotFound(String),
-
-    #[error("GitHub API rate limited: {0}")]
-    RateLimited(String),
-
-    #[error("GitHub API request failed: {0}")]
-    RequestFailed(String),
 }
